Ollama & Continue & Llama3/StarCode2 devfile (#2)
Added a devfile that starts a new workspace comprising containers based on the universal developer image and the ollama image. The Continue VS Code extension ist being installed after start-up and the models Llama3 and StarCoder2 (for autocomplete functionality) are being pulled using ollama. Additionally, the Continue extension is configured to use these two models as well as to disallow anonymous telemetry being sent. This yields a personal AI assistant for developers that ensures the inferencing is done entirely inside the cluster.